Moved chart into its own component
