+++ source/scss/app.scss
@@ -3,5 +3,5 @@
 @import "base/normalize";
 @import "base/breakpoints";
 @import "base/app";
-@import "base/menu";
-@import "base/home";
+@import "components/menu";
+@import "views/home";
